refactor: move styles types to type folder
