+++ src/app/components/button/button.tsx
@@ -1,5 +1,5 @@
 import { ButtonHTMLAttributes, forwardRef, MouseEvent } from 'react';
-import { ControlSize } from 'src/app/models/styles';
+import { ControlSize } from 'src/app/types/styles';
 import { Slot } from '@radix-ui/react-slot';
 import { StrictUnion } from 'src/app/types/union';
 import clsx from 'clsx';
