+++ src/app/components/icon-button/icon-button.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
 import { CrownSimple } from 'phosphor-react';
-import { ControlSizes } from 'src/app/models/styles';
+import { ControlSizes } from 'src/app/types/styles';
 import { ButtonColorVariants, ButtonStyleVariants } from '../button/button';
 import { IconButton } from './icon-button';
