+++ src/app/types/styles.ts
